※２点修正あります。
■テスト仕様書
------------------------------

ファイル名　　　　【　テスト仕様書　】
シート名　　　　　【  物件管理　】
バグＮｏ　　　　　【　64　】

------------------------------
■バグの内容
【物件登録】の画面に、削除ボタンが表示されている。
------------------------------
■修正完了後作業【エビデンス】
if文を追加。

------------------------------------------------------------------------------------------

■テスト仕様書
------------------------------

ファイル名　　　　【　テスト仕様書　】
シート名　　　　　【  物件管理　】
バグＮｏ　　　　　【　70　】

------------------------------
■バグの内容
【物件登録画面】で、「戻る」ボタンを押下すると、
FM画面へ遷移する。
------------------------------
■修正完了後作業【エビデンス】
valueの値を修正。

+++ SRC/libArticle.php
@@ -86,8 +86,8 @@ function subArticle()
 		return;
 	}
 	$sql = fnSqlArticleList(0, $sDel, $sArticle, $sRoom, $sKeyPlace, $sArticleNote, $sKeyBox, $sDrawing, $sSellCharge, $sPage, $orderBy, $orderTo);
-	$res = mysql_query($sql);
-	$row = mysql_fetch_array($res);
+	$res = mysqli_query($conn, $sql);
+	$row = mysqli_fetch_array($res);
 
 	$count = $row[0];
 
@@ -260,8 +260,10 @@ function subArticleEdit()
 		</table>
 
 		<a href="javascript:fnArticleEditCheck();"><img src="./images/<?php print $btnImage ?>" /></a>　
-		<a href="javascript:form.act.value='fManager';form.submit();"><img src="./images/btn_return.png" /></a>
-		&nbsp;&nbsp;<a href="javascript:fnArticleDeleteCheck(<?php print $articleNo ?>);"><img src="./images/btn_del.png" /></a>
+		<a href="javascript:form.act.value='articleSearch';form.submit();"><img src="./images/btn_return.png" /></a>
+		<?php if ($articleNo) { ?>
+			&nbsp;&nbsp;<a href="javascript:fnArticleDeleteCheck(<?php print $articleNo ?>);"><img src="./images/btn_del.png" /></a>
+		<?php } ?>
 	</form>
 <?php
 }
